Update and rename HidePlayer.cpp to HideOrShowPlayer.cpp
